cmd: remove `gen-manifests` in favor of the `images` version
This is a result of the outcome of
https://github.com/osbuild/osbuild-composer/pull/4124

Achilleas suggested that instead of having a copy of `gen-manifests`
in osbuild-composer we could use the one from images. AFAICT nothing
in the code references `gen-manifests` except the `test/README` so
it seems safe to remove and just use images.

The README got updated to point to `images` and we will also
need to remove `test/data/manifest` at a later point (after
checking that nothing else depends on this anymore).