test/dot-notation-aliases: override rhel-8 alias
The test relies on the existence of the /etc/dnf/vars/releasever file
and the value it specifies.  In SAP images, this file specifies the
locked release version for the system and it is written by the image
definition itself.  This makes it a good candidate for ensuring we are
using the correct distro version when building the image.

However, RHEL 8.10 SAP images don't lock the distro version because 8.10
is the last minor version in the RHEL 8 cycle.

Let's override the alias for RHEL 8 to point to 8.8, which is a version
that should contain a releasever.  The exact value we use isn't
important.  What we are testing here is the mechanism for resolving
aliases.