SPEC: take c10s and el10 into consideration for repo installation
Previously, the SPEC file was written in the way that c9s and el9 were
the latest major releases of the respective distributions and as such,
they could build images for all previous major versions of the same
distro. With c10s and el10, this is no longer true. Amend the section
which installs repository definitions.

The idea is that:
 - c10s / el10 are the latest major version and can build any previous
   major version of the same distro.
 - Any cs and el distro version can build images for the same major
   version.
 - c9s can build also c8s / el9 can build also el8 images.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>