tools: new script gen-manifests
A new script to replace ./cmd/gen-manifests.  The script queries the go
dependencies to get the version of osbuild/images that osbuild-composer
currently depends on, clones the osbuild/images repository at that
version, and uses the ./cmd/gen-manifests tool from there to generate
new manifests.

This is meant to replace ./cmd/gen-manifests so that we only have one
command and set of configurations for generating manifests, to avoid
drift between the two implementations.