test/api: silent dump_db()
When cleaning up api tests, the build job results are dumped from the
database to store in build-result.txt and are printed to the runner log
at the same time.  This makes it very difficult to read results as the
database dump prints very long lines that flood the job log.  The result
isn't really readable (or useful) unless the file is downloaded
separately.

Silence the `tee` command so that the file is created but no output is
printed from the db dump.