worker: add `endpoint()` helper that cannot fail and use it
Small commit to (hopefully) make the intent of the code more
clear when `serverURL.Parse()` is used. In most cases it is
used to generate an endpoint URL. As this is a realtive path
we do not need `Parse()` but can use `JoinPath()` instead which
has the nice benefit that it cannot error or panic.

It also contains some tests that I mostly added to be sure that
there are no regressions (test coverage is not great so there
is still a risk).