worker: remove `api.BasePath` and replace with MakeHTTPErrorHandler()
The `api.BasePath` is package global but the way it is used is that
multiple clients can (potentially) have different BasePath and
the way it is written right now could lead to confusing bugs.

This commit tweaks the `api` package so that `BasePath` does not
need to be set. The only place that this is needed is for the
`api.APIError()` which can just get the basePath as an argument.
With that we also need a helper to construct the api error handler
from the static HTTPErrorHandler to a dynamic MakeHTTPErrorHandler()
that will return a handler for the given basePath.

This should make https://github.com/osbuild/osbuild-composer/pull/4328
obsolete.