blueprint: port filesystem customization tests from images
Copied from
osbuild/images:pkg/blueprint/filesystem_customizations_test.go

These test features of the filesystem customizations:
- Deserialization
- Mountpoint policy validation

The filesystem customization _might_ be removed from osbuild/images and
osbuild-composer will be responsible for mapping the old customizations
onto the new.  Copying the tests over ensures that if and when we make
this change, the functionality of old format is still tested.