cloudapi: Add architecture support to the optional Blueprint
The blueprint may have the architecture set. If set it must match the
architecture set by the request. This ensures that clients have
correctly set the architecture and prevents unexpectedly ignoring the
blueprint architecture.

With the WELR API the blueprint architecture would override the host
arch. But with the Cloud API there is no host arch to override so the
request must have the expected architecture set as part of the request.

Related: RHEL-60125