tests/ubi-wsl: make cleanup best-effort
The cleanup section sometimes fail for no apparent reason despite the
actual test passing. This seems to be an issue with azure rather than
the test so this change enables the cleanup to fail. We have
cloud-cleaner anyway to remove any old resources. Opened COMPOSER-2371
to look into this later.