osbuild-composer: log NoReposLoadedError and continue
osbuild/images added an error type that's returned when the reporegistry
loader doesn't find any repository configurations to load [1].  This
lets callers decide whether to stop or continue execution based on
whether repository configurations are required.

Weldr (running on-prem) requires local repository configurations, while
the API (SaaS) doesn't.  However, at the time during startup when
reporegistry.LoadAllRepositories() is called, it is still not known
whether the Weldr API will be initialised.

Log the error message as an info message when it is detected but
continue execution.

[1] https://github.com/osbuild/images/pull/946