tools: use go toolbox instead of installing a binary
Repeat of 8554d6202dff0f887d0f8bec838f15c24af02e8d.

This is a much nicer (and newer) way of handling the Go version for the
project.  The issue that caused the previous revert was that the toolbox
version was being set to 1.22.0, which is older than what some
dependencies require, specifically osbuild/images, so the 'go mod tidy'
would downgrade images to match the required toolbox version.

Setting the go and toolbox version to 1.22.6 should make it all work as
expected.

This reverts commit 7e87d1e124c1bbd49f7a0afcd732fd83357f4e57.

Co-authored-by: Lukáš Zapletal <lukas@zapletalovi.com>