cloudapi/v2: minsize can be a string or integer
Define a reusable minsize property in the cloud API schema.
The minsize type is either a uint64 or a string with data units.

Internally, this creates an interface{} alias called Minsize and
optional uses of the property are of type *Minsize.  We define an
internal decodeMinsize() function to convert *Minsize to uint64.  This
is similar to the decodeSize() function in the blueprint, but made
specifically to handle *Minsize.

The API restricts the valid types for Minsize to just string or uint64,
but we also handle int in the decode function to make it easier to work
with by setting int literals:

    common.ToPtr(Minsize(2)) vs common.ToPtr(Minsize(uint64(2)))