osbuild-composer.spec: assume no modules in check step
Removes an ineffectual GO111MODULE=off assignment in the build phase, as
we require go to run in module mode at that stage.

In the check phase we run without go modules, so the same stipulation as
on RHEL applies, where the vendor directory gets ignored in certain
cases. Also remove `-mod=vendor` from GOFLAGS on RHEL during the check
phase, as without go modules this is an ineffectual flag.