CI: trigger EDGE CI from GH action, not GitLab
The way we trigger EDGE CI is silly. We boot a runner in GitLab just to
use CURL to trigger EDGE CI pipeline on GitHub. None of this is really
needed and we should be triggering EDGE CI from GH Action.

Subsequently, the EDGE CI pipeline expects a PR number to be triggered,
but we have been triggering also for merges to `main`. It does not show
up in osbuild-composer runs on `main`, but in reality, the pipeline gets
triggered and always fails. This rework tries to mitigate the situation
by not triggering the pipeline if the workflow is not run as part of PR
to the `main` branch.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>