Update and reduce testing to RHEL 8.10
Reduce test configurations to only run on and for RHEL 8.10 and bump the
terraform commit ID.
Update all GitHub runners to Ubuntu 22.04 (from 20.04, deprecated).

Install dnf4 in the Python Lint job for dnf-json since it runs on Fedora
41 now.

Update the Azure test cases to fix the instance definition and bring in
any other test updates.