Update and configure tests for RHEL 8.10 & RHEL 9.4
Reduce test configurations to only run on and for RHEL 8.10 and RHEL
9.4 and bump the terraform commit ID.
Some tests were already reduced to 9.4 on the branch and now have 8.10
runners added.

Install dnf4 in the Python Lint job for dnf-json since it runs on Fedora
41 now.

Update the Azure test cases to fix the instance definition and bring in
any other test updates.