Repositories: delete RHEL-8 EOL release repositories
For RHEL-10, we have a special version of RHEL-8 repositories, that do
not contain the auxiliary key. This is due to the fact that the key uses
SHA-1, which is not allowed by default by RHEL-10 crypto policy.

We deleted repositories for these releases in osbuild/images, but we
need to delete them also here, to ensure that they don't end up in the
RPMs and that the cross-distro test case does not fail.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>