cloudapi: Add support for depsolving a blueprint + image type
This uses the blueprint and the image type's manifest with a minimal
imageRequest to get the package sets to use for depsolving, instead of
just using the packages in the blueprint.

This does not work with iot image types that require an ostree url.
Depsolving iot commits doesn't make sense anyway, since the blueprint
packages have no effect on it.

Includes simple tests for depsolving with image type, and returning an
error if the image type isn't recognized.

Related: RHEL-60125