Test/regression-old-worker-new-composer: make repo construction robust
The test case needs to install RPMs for an osbuild-composer version
which is older by three releases, than the version under test. This is
done by looking up the GH tag and its corresponding commit SHA. However,
it may happen that the CI pipeline for that commit SHA failed and there
were no RPMs built that can be installed.

To make the test case more robust, enhance it, so that it can iterate 10
commits back to the past, in order to find a commit SHA with a working
corresponding repository.

There is a slight chance that the traversal to older commits may cross
another release tag, so the RPM repository may contain older
osbuild-composer worker version than desired. In such case the test case
will fail. To not complicate the test case further, let's keep this
scenario unhandled, but explicitly mentioned in a comment.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>