Test/regression-old-worker-new-composer: don't setup composer repo twice
The test case is testing that old version of osbuild-composer worker
still works with the current version of the osbuild-composer under test.
This is to ensure that we do not break backward compatibility of the
worker API.

The test case first installs the version of osbuild-composer under test,
to determine its version. It uses GH to determine the upstream commit
SHA of an osbuild-composer version older by three releases than the
current version under test. The installed osbuild-composer* RPMs are
then removed.

In the next phase, the test case sets up RPM repo for the "old" composer
version from CI S3 cache and installs the osbuild-composer-worker
package from it. Lastly, the current version of osbuild-composer under
test is pulled as a container, which is built by a previous job within
the same CI pipeline.

Finally, the test case runs a testing image build for qcow2 type, using
the current version of osbuild-composer under test running inside a
container, and the old worker running locally. The built image is then
downloaded from S3 and inspected using osbuild-image-info tool.

Previously, the test case used to setup osbuild-composer RPM repo a
second time, without any apparent reason. The second repo used a third
version of osbuild-composer (different than the one used for the old
worker and the current version under test), which was creating another
dependency on old CI pipeline runs. This was most probably a leftover
from previous versions of the test case.

Let's remove the second setup of osbuild-composer RPM repo to reduce the
situations that can cause the test case to fail, in case the old CI
pipelines failed or didn't run at all.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>