Dockerfile-ubi-packer: do not use quay.io/app-sre/packer image
The quay.io/app-sre/packer:latest is outdated and has old Packer version
installed. Since we are the only user of it, AppSRE suggested that we
build our own image.

Change the containerfile to be built on top of UBI9 (the AppSRE image
was based on UBI8) and install a subset of packages that were previously
installed in the quay.io/app-sre/packer [0].

Specifically, install:
 - Packer
 - Ansible (since we use it to modify images)
 - awscli (to interact with AWS)

[0] https://github.com/app-sre/container-images/blob/8f0dff56739fd270cbb5d07530ae0b90e23174a6/logrotate/Dockerfile

Signed-off-by: Tomáš Hozza <thozza@redhat.com>