test/container-embedding: adjust the test for multi-arch container
We had to push a new multi-arch container build to the
fedora-minimal:latest tag, which broke the test case. The test case
previously hard-coded various digests and IDs of container images. To
save ourselves some future work, modify the test case to determine all
the values dynamically, so that we don't need to adjust them each time
we update the container image.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>