templates/packer: add vector forwarding from executor to worker
This configures forwarding from the executor to the worker to
cloudwatch.

This prepares for removing internet access from the executor. Vector
forwarding via the worker is easier than trying to get the executor
cloudwatch access through VPC endpoints. The latter would make the
secure instance code a lot more complex.